Fiesta - First Merge

+++ BodaKaoz/LunaMiel/models.py
@@ -14,7 +14,7 @@ class Actividad(models.Model):
 	Plan=models.ForeignKey(Plan, on_delete=models.CASCADE)
 	nombre=models.CharField(max_length=50)
 	imagen=models.ImageField(null=True, blank=True, default=None)
-	precio=models.BigIntegerField(default=0)
+	precio=models.IntegerField(default=0)
 	class Meta:
 		verbose_name = "Actividad"
 		verbose_name_plural = "Actividades"
@@ -24,7 +24,7 @@ class Hotel(models.Model):
 	descripcion=models.CharField(max_length=50)
 	calificacion=models.IntegerField()
 	imagen=models.ImageField(null=True, blank=True, default=None)
-	precio=models.BigIntegerField(default=0)
+	precio=models.IntegerField(default=0)
 	class Meta:
 		verbose_name = "Hotel"
 		verbose_name_plural = "Hoteles"
@@ -45,7 +45,7 @@ class HotelPlan(models.Model):
 
 class LunaMielEvento(models.Model):
 	Boda=models.ForeignKey(Boda, on_delete=models.CASCADE)
-	precio=models.BigIntegerField(default=0)
+	precio=models.IntegerField(default=0)
 		
 class ActividadCarrito(models.Model):
 	LunaMielEvento=models.ForeignKey(LunaMielEvento, on_delete=models.CASCADE)
